requirements.txt
allure_python_commons==2.9.45
pytest==6.2.5
requests==2.27.1
xlrd==2.0.1
